pyproject.toml
[tool.poetry]
name = "lang-pz3"
version = "0.1.0"
description = "LangSmith/LangChain workflow implementation"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
langgraph = ">=0.1.0"
langchain = ">=0.3.0"
langchain-openai = ">=0.3.12"
langsmith = ">=0.3.0"
pydantic = ">=2.0,<3.0"
python-dotenv = ">=1.0.0"
requests = ">=2.31.0"
openai = ">=1.0.0"
multidict = "==6.0.4"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"